made day14 conflict free
